Automatic Trailing Slash Compatibility
When checking if a redirect is present, an exact match is currently required (with or without the trailing slash). This change faciliates trailing slash agnostic support - a redirect for path /test is the same as /test/. Query strings are also supported in this manner.

Unit Tests Passed: OK (9 tests, 18 assertions). Also cURL checks performed to confirm as well.